Update Perf return types
 - Neither function should ever fail, and thus they return unsigned
   values.

+++ libtock/perf.c
@@ -4,10 +4,10 @@ int perf_count(void) {
   return command(DRIVER_PERF, 0, 0, 0);
 }
 
-int perf_cycles(void) {
-  return command(DRIVER_PERF, 1, 0, 0);
+unsigned perf_cycles(void) {
+  return (unsigned)command(DRIVER_PERF, 1, 0, 0);
 }
 
-int perf_instructions_retired(void) {
-  return command(DRIVER_PERF, 2, 0, 0);
+unsigned perf_instructions_retired(void) {
+  return (unsigned)command(DRIVER_PERF, 2, 0, 0);
 }
